requirements.txt
# =============================================
# Core Libraries
# =============================================
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0
scikit-learn>=1.0.0

# =============================================
# Deep Learning
# =============================================
tensorflow>=2.13.0
tensorflow-federated>=0.55.0
tensorflow-privacy>=0.8.0
torch>=2.0.0
torchvision>=0.15.0

# =============================================
# Explainable AI
# =============================================
lime>=0.2.0.1
shap>=0.43.0
tf-explain>=0.3.1
captum>=0.6.0

# =============================================
# Federated Learning & Privacy
# =============================================
syft>=0.8.0
opacus>=1.4.0
crypten>=0.4.0

# =============================================
# Image Processing
# =============================================
opencv-python>=4.5.0
Pillow>=8.3.0
albumentations>=1.3.0

# =============================================
# Data Science & ML
# =============================================
scipy>=1.7.0
jupyter>=1.0.0
ipykernel>=6.0.0
notebook>=6.4.0

# =============================================
# Utilities
# =============================================
tqdm>=4.62.0
pyyaml>=5.4.0
python-dotenv>=0.19.0
click>=8.0.0

# =============================================
# Testing
# =============================================
pytest>=6.2.0
pytest-cov>=2.12.0

# =============================================
# Development
# =============================================
black>=21.9.0
flake8>=3.9.0
pre-commit>=2.15.0